Split Monitor into Partial and RecoveryMonitor
Separate trigger-only monitors from recovery/finalisation logic. Introduce RecoveryMonitor for types that need to complete stale records on next process start and keep PartialMonitor for runtime triggers. Make LaunchObject a PartialMonitor (its trigger remains) and move its completion logic to a RecoveryMonitor extension file; do the same for SessionObject. Remove LaunchObject.complete from the background action table. Add and rename tests to cover Launch monitor behavior and Session/Launch recovery logic.
