+++ Sources/Scout/Core/Lifecycle/Session/SessionObject+Recovery.swift
@@ -7,7 +7,7 @@
 
 import CoreData
 
-extension SessionObject {
+extension SessionObject: RecoveryMonitor {
     /// Closes sessions from previous launches that were not properly
     /// completed — typically because the app crashed.
     ///
